+++ src/vmc/vmc_data.h
@@ -51,44 +51,44 @@ class AdminCash: public vmcData<uint32_t> {
 
 /* Tariff stored in EEPROM: */
 
-class tariff: public vmcData<double> {
+class tariff: public vmcData<float> {
     public:
         static tariff * get_default_instance();
-        virtual vmc_data_error_t set(double value);
+        virtual vmc_data_error_t set(float value);
         virtual bool load();
-        static double DEFAULT_TARIFF;
+        static float DEFAULT_TARIFF;
 };
 
 
 /* Calibration stored in EEPROM: */
 
-class Calibration: public vmcData<double> {
+class Calibration: public vmcData<float> {
     public:
         static Calibration * get_default_instance();
-        virtual vmc_data_error_t set(double value);
+        virtual vmc_data_error_t set(float value);
         virtual bool load();
-        static double DEFAULT_CALIBRATION;
+        static float DEFAULT_CALIBRATION;
 };
 
 /* Flow calculation stored in EEPROM: */
 
-class FlowCalculation: public vmcData<double> {
+class FlowCalculation: public vmcData<float> {
     public:
         static FlowCalculation * get_default_instance();
-        virtual vmc_data_error_t set(double value);
+        virtual vmc_data_error_t set(float value);
         virtual bool load();
-        static double DEFAULT_FLOW_CALCULATION;
+        static float DEFAULT_FLOW_CALCULATION;
 };
 
 /* Quantities struct: */
 typedef struct quantities
 {
-    double quantity1;
-    double quantity2;
-    double quantity3;
-    double quantity4;
-    double quantity5;
-    double quantity6;
+    float quantity1;
+    float quantity2;
+    float quantity3;
+    float quantity4;
+    float quantity5;
+    float quantity6;
 
 }quantities_t;
 
@@ -98,10 +98,10 @@ class Quantities: public vmcData<quantities_t> {
         static Quantities * get_default_instance();
         virtual vmc_data_error_t set(quantities_t value);
         virtual bool load();
-        static double DEFAULT_QUANTITTY_1;
-        static double DEFAULT_QUANTITTY_2;
-        static double DEFAULT_QUANTITTY_3;
-        static double DEFAULT_QUANTITTY_4;
-        static double DEFAULT_QUANTITTY_5;
-        static double DEFAULT_QUANTITTY_6;
+        static float DEFAULT_QUANTITTY_1;
+        static float DEFAULT_QUANTITTY_2;
+        static float DEFAULT_QUANTITTY_3;
+        static float DEFAULT_QUANTITTY_4;
+        static float DEFAULT_QUANTITTY_5;
+        static float DEFAULT_QUANTITTY_6;
 };
